+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 **Added**
 
 - When using the [lint](./README.md#lint) extension for the JavaScript mode, global variables will be highlighted based on the [ESLint legacy configuration](https://eslint.org/docs/latest/use/configure/language-options-deprecated#specifying-environments)
+- The [signatureHelp](./README.md#signaturehelp) extension now supports dismissing the tooltip with the `Escape` key
 
 ## 3.10.2